refactor: rename console module to command
Renamed `src/console/` to `src/command/` to better reflect the module's purpose of handling terminal commands for bot administration.

- Moved src/console/mod.rs → src/command/mod.rs
- Updated module import in main.rs
